+++ book-lib/src/App.js
@@ -1,17 +1,17 @@
-import "./App.css";
+import "./index.css";
 import { Route, Routes } from "react-router-dom";
-import SearchBooks from "./components/Search";
-import Navbar from "./components/Navbar";
+import SearchBooks from "./components/Search/Search";
+import Navbar from "./components/Navbar/Navbar";
 
 function App() {
-  return (
-    <>
-      <Navbar />
-      <Routes>
-        <Route path="/search" element={<SearchBooks />} />
-      </Routes>
-    </>
-  );
+	return (
+		<>
+			<Navbar />
+			<Routes>
+				<Route path='/search' element={<SearchBooks />} />
+			</Routes>
+		</>
+	);
 }
 
 export default App;
