--- book-lib/src/Styles/styles.css
@@ -1,88 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Titillium+Web:wght@300;400;700&display=swap");
-
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  font-family: "Titillium Web", sans-serif;
-}
-
-:root {
-  --mainColor: #deb887;
-  --mainColorLight: #5767aa;
-  --secondaryColor: #db2b2b;
-  --textColor: #eee;
-}
-
-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  height: 80px;
-  background-color: var(--mainColor);
-  color: var(--textColor);
-}
-
-nav a {
-  margin: 0 1rem;
-  color: var(--textColor);
-  text-decoration: none;
-}
-
-nav a:hover {
-  color: var(--secondaryColor);
-}
-
-header .nav-btn {
-  padding: 5px;
-  cursor: pointer;
-  background: transparent;
-  border: none;
-  outline: none;
-  color: var(--textColor);
-  visibility: hidden;
-  opacity: 0;
-  font-size: 1.8rem;
-}
-
-header div,
-nav {
-  display: flex;
-  align-items: center;
-}
-
-@media only screen and (max-width: 1024px) {
-  header .nav-btn {
-    visibility: visible;
-    opacity: 1;
-  }
-
-  header nav {
-    position: fixed;
-    top: -100vh;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 1.5rem;
-    background-color: var(--mainColor);
-    transition: 1s;
-  }
-
-  header .responsive_nav {
-    transform: translateY(100vh);
-  }
-
-  nav .nav-close-btn {
-    position: absolute;
-    top: 2rem;
-    right: 2rem;
-  }
-
-  nav a {
-    font-size: 1.5rem;
-  }
-}
